+++ packages/headless/src/modules/customer-auth/rest/dto/getCustomer.dto.ts
