+++ packages/headless/src/modules/customer-auth/rest/dto/register.dto.ts
