+++ packages/headless/src/modules/customer/rest/dto/getCustomer.dto.ts
